src/app/app.component.spec.ts
import { TestBed,ComponentFixture } from '@angular/core/testing';
import { AppComponent } from './app.component';
import { HttpClientModule } from '@angular/common/http';
import { ProfileComponent } from './profile/profile.component';
import { RepoComponent } from './repo/repo.component';
// import { FormGroup, FormControl } from '@angular/forms';
import { FormsModule, ReactiveFormsModule } from '@angular/forms';

describe('AppComponent', () => {
  let component: AppComponent;
  let fixture: ComponentFixture<AppComponent>;

  beforeEach(async() => {
    await TestBed.configureTestingModule({
    declarations: [
      AppComponent,
      ProfileComponent,
      RepoComponent,
      
    ],
    imports: [ HttpClientModule,ReactiveFormsModule,FormsModule ],
  }).compileComponents();
  fixture = TestBed.createComponent(AppComponent);
  component = fixture.componentInstance;
});

  it('should create the app', () => { 
    expect(component).toBeTruthy();
  });

  it(`should have as title 'fyle-frontend-challenge'`, () => {
    
    expect(component.title).toEqual('fyle-frontend-challenge');
  });

  it('should render title', () => {
    fixture.detectChanges();
    const compiled = fixture.nativeElement as HTMLElement;
    expect(compiled.querySelector('.content span')?.textContent).toContain('fyle-frontend-challenge');
  });
});
